Atomization of HarvestDetail component
